+++ lib/src/widgets/fields/ds_text_form_field.widget.dart
@@ -1,6 +1,9 @@
-import 'package:blip_ds/blip_ds.dart';
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
+
+import '../../themes/colors/ds_colors.theme.dart';
+import '../../themes/texts/styles/ds_body_text_style.theme.dart';
+import '../../utils/ds_utils.util.dart';
+import 'ds_input_decoration.widget.dart';
 
 class DSTextFormField extends StatelessWidget {
   DSTextFormField({
@@ -13,8 +16,8 @@ class DSTextFormField extends StatelessWidget {
     this.textCapitalization = TextCapitalization.sentences,
     this.showEmojiButton = false,
     this.obscureText = false,
-    this.isEnabled = true,
     this.focusNode,
+    this.isEnabled = true,
   });
 
   final TextEditingController controller;
@@ -25,92 +28,79 @@ class DSTextFormField extends StatelessWidget {
   final TextCapitalization textCapitalization;
   final bool showEmojiButton;
   final bool obscureText;
-  final bool isEnabled;
   final FocusNode? focusNode;
+  final bool isEnabled;
 
-  final hasFocus = false.obs;
   final _scrollController = ScrollController();
 
   @override
-  Widget build(BuildContext context) => Focus(
-        onFocusChange: hasFocus,
-        child: Obx(
-          () => AnimatedContainer(
-            duration: DSUtils.defaultAnimationDuration,
-            padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 9),
-            decoration: BoxDecoration(
-              borderRadius: const BorderRadius.all(
-                Radius.circular(25),
-              ),
-              border: Border.all(
-                width: 1,
-                color: hasFocus.value
-                    ? DSColors.primaryMain
-                    : isEnabled
-                        ? DSColors.neutralMediumSilver
-                        : DSColors.neutralLightBox,
-              ),
-              color: isEnabled
-                  ? DSColors.neutralLightSnow
-                  : DSColors.neutralLightWhisper,
+  Widget build(BuildContext context) => DSInputDecoration(
+        padding: const EdgeInsets.only(
+          left: 12.0,
+          right: 6.0,
+          top: 10.0,
+          bottom: 10.0,
+        ),
+        child: Scrollbar(
+          controller: _scrollController,
+          radius: const Radius.circular(5),
+          child: Padding(
+            padding: const EdgeInsets.only(
+              right: 6.0,
             ),
-            child: Scrollbar(
-              controller: _scrollController,
-              radius: const Radius.circular(5),
-              child: Row(
-                crossAxisAlignment: CrossAxisAlignment.end,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Flexible(
-                    child: ConstrainedBox(
-                      constraints: const BoxConstraints(
-                        maxHeight: 100.0,
-                      ),
-                      child: TextFormField(
-                        controller: controller,
-                        scrollController: _scrollController,
-                        enabled: isEnabled,
-                        obscureText: obscureText,
-                        style: const DSBodyTextStyle(),
-                        decoration: InputDecoration(
-                          isDense: true,
-                          contentPadding: EdgeInsets.zero,
-                          border: InputBorder.none,
-                          focusedBorder: InputBorder.none,
-                          enabledBorder: InputBorder.none,
-                          errorBorder: InputBorder.none,
-                          disabledBorder: InputBorder.none,
-                          hintText: hint,
-                          hintStyle: const DSBodyTextStyle(
-                            color: DSColors.neutralMediumElephant,
-                          ),
+            child: Row(
+              crossAxisAlignment: CrossAxisAlignment.end,
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                Flexible(
+                  child: ConstrainedBox(
+                    constraints: const BoxConstraints(
+                      maxHeight: 100.0,
+                    ),
+                    child: TextFormField(
+                      controller: controller,
+                      scrollController: _scrollController,
+                      enabled: isEnabled,
+                      obscureText: obscureText,
+                      style: const DSBodyTextStyle(),
+                      decoration: InputDecoration(
+                        isDense: true,
+                        contentPadding: EdgeInsets.zero,
+                        border: InputBorder.none,
+                        focusedBorder: InputBorder.none,
+                        enabledBorder: InputBorder.none,
+                        errorBorder: InputBorder.none,
+                        disabledBorder: InputBorder.none,
+                        hintText: hint,
+                        hintStyle: const DSBodyTextStyle(
+                          color: DSColors.neutralMediumSilver,
                         ),
-                        cursorColor: DSColors.primaryMain,
-                        cursorHeight: 20.0,
-                        textCapitalization: textCapitalization,
-                        textInputAction: textInputAction,
-                        maxLines: obscureText ? 1 : maxLines,
-                        onChanged: onChanged,
-                        focusNode: focusNode,
                       ),
+                      cursorColor: DSColors.primaryMain,
+                      cursorHeight: 20.0,
+                      textCapitalization: textCapitalization,
+                      textInputAction: textInputAction,
+                      maxLines: obscureText ? 1 : maxLines,
+                      onChanged: onChanged,
+                      focusNode: focusNode,
                     ),
                   ),
-                  if (showEmojiButton)
-                    IconButton(
-                      /// TODO: Implement to open emoji keyboard!!
-                      onPressed: () {},
-                      padding: EdgeInsets.zero,
-                      splashRadius: 15,
-                      constraints: const BoxConstraints(),
-                      icon: Image.asset(
-                        'assets/images/emoji.png',
-                        package: DSUtils.packageName,
-                        height: 20.5,
-                        width: 20.5,
-                      ),
-                    )
-                ],
-              ),
+                ),
+                if (showEmojiButton)
+                  IconButton(
+                    /// TODO: Implement to open emoji keyboard!!
+                    onPressed: () {},
+                    padding: EdgeInsets.zero,
+                    splashRadius: 15,
+                    constraints: const BoxConstraints(),
+                    icon: Image.asset(
+                      'assets/images/emoji.png',
+                      package: DSUtils.packageName,
+                      height: 20.5,
+                      width: 20.5,
+                    ),
+                  )
+              ],
             ),
           ),
         ),