+++ lib/src/widgets/texts/ds_headline_large_text.widget.dart
@@ -3,36 +3,44 @@ import 'ds_text.widget.dart';
 
 /// A Design System's [Text] primarily used by large titles.
 ///
-/// Sets [DSHeadlineLargeTextStyle] as [style] default value. This style's font variant is $fs-20-h2.
+/// Sets [DSHeadlineLargeTextStyle] as [style] default value. This style's font variant is $fs-20-h4.
 class DSHeadlineLargeText extends DSText {
-  /// Creates a Design System's [Text] with $fs-20-h2 font variant.
+  /// Creates a Design System's [Text] with $fs-20-h4 font variant.
   DSHeadlineLargeText(
     super.text, {
     super.key,
+    super.color,
     super.linkColor,
     super.overflow,
     super.textAlign,
     super.maxLines,
+    super.fontWeight,
     super.shouldLinkify,
     super.isSelectable,
   }) : super(
           style: DSHeadlineLargeTextStyle(
+            color: color,
             overflow: overflow,
+            fontWeight: fontWeight,
           ),
         );
 
   DSHeadlineLargeText.rich(
     super.textSpan, {
     super.key,
+    super.color,
     super.linkColor,
     super.overflow,
     super.textAlign,
     super.maxLines,
+    super.fontWeight,
     super.shouldLinkify,
     super.isSelectable,
   }) : super.rich(
           style: DSHeadlineLargeTextStyle(
+            color: color,
             overflow: overflow,
+            fontWeight: fontWeight,
           ),
         );
 }